fix(ft8): expose FT8/FT4/WSPR in the toolbar MODE picker too
The toolbar MODE favorites dropdown (the always-visible mode grid) didn't list
the digital modes — only the Mode tile did. Add FT8/FT4/WSPR to the dropdown
and branch onSelect to enterDigital for them. Factor enterDigital into a shared
state/enter-digital module (no import cycle) used by both pickers.
